% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best.linmod.R
\name{best.linmod}
\alias{best.linmod}
\title{best.linmod}
\usage{
best.linmod(
  data,
  response_var,
  num_predictors = length(data) - 1,
  train_size = 0.8
)
}
\arguments{
\item{data}{dataset}

\item{response_var}{the dependent variable}

\item{num_predictors}{number of predictor variables, default is number of variables -1}

\item{train_size}{percentage of the data that you want to allocate to the training set, default 0.8}
}
\value{
A summary including the dependent variable, the predictor variables,
and the percentage of variation in the dependent variable that the predictor
variables account for.
}
\description{
best.lm is a function to create the best possible linear model with a
specified number of predictor variables. It splits the data into a training
and test set, and the sizes of the sets can be changed in the "train_size"
parameter. Then it iterates through every possible linear model and finds the
best one based on the adjusted R squared and mean squared error.
}
\examples{

best.linmod(mtcars, "mpg")
best.linmod(mtcars, "mpg", 4, 0.7)

}
